Feat: 컴포넌트, svg 분리
